Split the "main.ts" file into several under same namespace
