+++ index.html
@@ -12,7 +12,7 @@
 
     <link rel="icon" href="data:,">
     <link rel="stylesheet" href="style.css" />
-    <script src="dist/main.js" defer></script>
+    <script src="dist/bundle.js" defer></script>
 
 </head>
