+++ src/entity/project.ts
@@ -1,0 +1,16 @@
+namespace App {
+  export enum ProjectStatus {
+    Active = "active",
+    Finished = "finished",
+  }
+
+  export class Project {
+    constructor(
+      public id: string,
+      public title: string,
+      public description: string,
+      public peopleCount: number,
+      public status: ProjectStatus
+    ) {}
+  }
+}
