+++ tsconfig.json
@@ -5,10 +5,11 @@
     "target": "ESNext" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
     "experimentalDecorators": true /* Enable experimental support for legacy experimental decorators. */,
     /* Modules */
-    "module": "commonjs" /* Specify what module code is generated. */,
+    "module": "AMD" /* Specify what module code is generated. */,
     "rootDir": "./src" /* Specify the root folder within your source files. */,
     /* Emit */
     "sourceMap": true /* Create source map files for emitted JavaScript files. */,
+    "outFile": "./dist/bundle.js" /* Specify a file that bundles all outputs into one JavaScript file. If 'declaration' is true, also designates a file that bundles all .d.ts output. */,
     "outDir": "./dist" /* Specify an output folder for all emitted files. */,
     "removeComments": true /* Disable emitting comments. */,
     "noEmitOnError": true /* Disable emitting files if any type checking errors are reported. */,
